% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction_wrapper.R
\name{dca}
\alias{dca}
\title{Deep count autoencoder}
\usage{
dca(exprDat, ...)
}
\arguments{
\item{exprDat}{Unnomalized, unscaled gene expression matrix, with cell names as rows and gene names as columns}

\item{...}{Additional arguments to pass to dca}
}
\description{
Performs denoising of gene expression using dca (https://github.com/theislab/dca)
}
