% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{dim.from.python}
\alias{dim.from.python}
\title{dim.from.python}
\usage{
\method{dim}{from.python}(seuratObj, python.dataframe, reduction.save)
}
\arguments{
\item{python.dataframe}{Dataframe returned by a Python function}

\item{reduction.save}{Name to use for the reduction (i. e. tsne, umap, etc...)}
}
\value{
A Seurat object with the dataframe stored in the seuratObj@dr$reduction.use slot
}
\description{
Helper function to assist entering dimensional reduction data from Python reduction methods
}
