% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{DoDCA}
\alias{DoDCA}
\title{DoDCA}
\usage{
DoDCA(seuratObj, ...)
}
\arguments{
\item{...}{Extra parameters to pass to the dca function.}
}
\value{
A Seurat object with imputed data stored in seuratObj@dr$dca slot
}
\description{
Use deep count autoencoder to impute data
}
