% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{DofastTSNE}
\alias{DofastTSNE}
\title{DofastTSNE}
\usage{
DofastTSNE(seuratObj, reduction.use = "pca",
  reduction.save = "fasttsne", ...)
}
\arguments{
\item{reduction.use}{Prior dimensional reduction to use for calculations
(i.e. pca, ica, cca, etc...). Default: pca}

\item{reduction.save}{Name to use for the reduction (i. e. tsne, umap,
etc...). Default: fastTSNE}

\item{...}{Extra parameters to pass to the fastTSNE function.}
}
\value{
Seurat object with the tSNE projection stored in the
  seuratObj@dr$tsne slot (unless otherwise specified)
}
\description{
Do tSNE projection on a Seurat object using the fastTSNE
library, with FIt-SNE selected by default
}
