% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{DoopenTSNE}
\alias{DoopenTSNE}
\title{DoopenTSNE}
\usage{
DoopenTSNE(object, reduction.use = "pca", reduction.save = "openTSNE",
  ...)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object to be transformed.}

\item{reduction.use}{Prior dimensional reduction to use for calculations
(i.e. pca, ica, cca, etc...). Default: pca}

\item{reduction.save}{Name to use for the reduction (i. e. tsne, umap,
etc...). Default: openTSNE}

\item{...}{Extra parameters to pass to the openTSNE function.}
}
\description{
Perform tSNE projection on a Seurat object using the openTSNE
library, with FIt-SNE selected by default
}
