% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceFunction.R
\name{DoPaCMAP}
\alias{DoPaCMAP}
\title{DoPaCMAP}
\usage{
DoPaCMAP(object, assay_use = "RNA", reduction_save = "PAC", ...)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object to be transformed.}

\item{assay_use}{Assay on which to perform dimensional reduction}

\item{reduction_save}{Name to use for the reduction (i. e. tsne, umap,
etc...). Default: PAC}

\item{...}{Extra parameters to pass to the PaCMAP function.}
}
\description{
Perform PaCMAP projection on a Seurat object using the pacmap
library
}
