% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{DoPHATE}
\alias{DoPHATE}
\title{DoPHATE}
\usage{
DoPHATE(seuratObj, reduction.use = "pca", reduction.save = "phate",
  ...)
}
\arguments{
\item{reduction.use}{Prior dimensional reduction to use for calculations
(i.e. pca, ica, cca, etc...). Default: pca}

\item{reduction.save}{Name to use for the reduction (i. e. tsne, umap,
etc...). Default: phate}

\item{...}{Extra parameters to pass to the phate function.}
}
\value{
Seurat object with the PHATE projection stored in the
  seuratObj@dr$phate slot (unless otherwise specified)
}
\description{
Project trajectory-based dimensional reduction using PHATE
}
