% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceatlas2.R
\name{fa2}
\alias{fa2}
\title{ForceAtlas2

An R wrapper around the Python implementation of ForceAtlas2 found at
https://github.com/bhargavchippada/forceatlas2

Currently a very restricted version}
\usage{
fa2(init_pos, adjacencies, iterations = 500, ...)
}
\arguments{
\item{init_pos}{Array of initial positions, i.e. PCA or PAGA embeddings}

\item{adjacencies}{A graph in 2D format}

\item{iterations}{Number of times to iterate the main loop}

\item{...}{Additional arguments to pass to the forceatlas2 object instantiation}
}
\description{
ForceAtlas2

An R wrapper around the Python implementation of ForceAtlas2 found at
https://github.com/bhargavchippada/forceatlas2

Currently a very restricted version
}
