% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{GoDCA}
\alias{GoDCA}
\title{GoDCA}
\usage{
GoDCA(seuratObj, ...)
}
\arguments{
\item{...}{Extra parameters to pass to the phenograph function.}

\item{prefix}{String prefix to used as in the column name entered in the meta.data slot}
}
\value{
A Seurat object with imputed data stored in seuratObj@dr$dca slot
}
\description{
Use deep count autoencoder to impute data
}
