% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{GofitSNE}
\alias{GofitSNE}
\title{GofitSNE}
\usage{
GofitSNE(seuratObj, reduction.use = "pca", reduction.save = "fitsne",
  ...)
}
\arguments{
\item{reduction.use}{Prior dimensional reduction to use for calculations (i.e. pca, ica, cca, etc...). (default: pca)}

\item{reduction.save}{Name to use for the reduction (i. e. tsne, umap, etc...). (default: fitsne)}

\item{...}{Extra parameters to pass to the FIt-SNE function.}
}
\value{
Seurat object with the tSNE projection stored in the seuratObj@dr$tsne slot (unless otherwise specified)
}
\description{
Run tSNE projection on a Seurat object using the FIt-SNE function
}
