% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction_wrapper.R
\name{multicoreTSNE}
\alias{multicoreTSNE}
\title{multicoreTSNE}
\usage{
multicoreTSNE(r.data.frame, n.components = 3, perplexity = 30,
  early.exaggeration = 12, learning.rate = 200, n.iter = 2000,
  n.iter.without.progress = 30, min.grad.norm = 1e-07,
  metric = "euclidean", init = "random", verbose = 1,
  random.state = NULL, method = "barnes_hut", angle = 0.5,
  n.jobs = NULL, cheat.metric = TRUE)
}
\arguments{
\item{cheat_metric}{}
}
\value{
dataframe
}
\description{
An R wrapper for the Multicore-TSNE Python library found at https://github.com/DmitryUlyanov/Multicore-TSNE
}
