% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paga.R
\name{PAGAplot}
\alias{PAGAplot}
\title{PAGAplot}
\usage{
PAGAplot(object, edge_scale_weight = 0.2)
}
\arguments{
\item{object}{Seurat object with PAGA in misc slot to plot}

\item{edge_scale_weight}{Factor to scale edge line segment weight by.  Default: 0.5}
}
\description{
Plot the results from PAGA
}
