% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{PushData}
\alias{PushData}
\alias{PushData.Seurat}
\alias{PushData.seurat}
\title{PushData}
\usage{
PushData(object, ...)

\method{PushData}{Seurat}(seuratObj, python.dataframe, reduction.save,
  assay.used)

\method{PushData}{seurat}(seuratObj, python.dataframe, reduction.save)
}
\arguments{
\item{python.dataframe}{Dataframe returned by a Python function}

\item{reduction.save}{Name to use for the reduction (i. e. tsne, umap,
etc...)}

\item{assay.used}{Assay from which the data that is dimensionally reduced comes}
}
\value{
A Seurat object with the dataframe stored in the
  seuratObj@dr$reduction.use slot
}
\description{
Helper function to assist entering dimensional reduction data from Python
reduction methods
}
