% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{python.dim.reduction.bridge}
\alias{python.dim.reduction.bridge}
\title{python.dim.reduction.bridge}
\usage{
python.dim.reduction.bridge(seuratObj, reduction.use, reduction.save,
  function.use, ...)
}
\arguments{
\item{reduction.use}{Prior dimensional reduction to use for calculations (i.e. pca, ica, cca, etc...)}

\item{reduction.save}{Name to use for the reduction (i. e. tsne, umap, etc...)}

\item{function.use}{Dimensional reduction function to call.}

\item{...}{Extra parameters to pass to the dimensional reduction function.}
}
\description{
Generalized helper function that pulls the data from a Seurat object, passes the dataframe to a Python function
and places the resulting dataframe in the appropriate slot
}
