% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceFunction.R
\name{DooptSNE}
\alias{DooptSNE}
\title{DooptSNE}
\usage{
DooptSNE(object, reduction.use = "pca", reduction.save = "optsne", ...)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object to be transformed.}

\item{reduction.use}{Prior dimensional reduction to use for calculations
(i.e. pca, ica, cca, etc...). Default: pca}

\item{reduction.save}{Name to use for the reduction (i. e. tsne, umap,
etc...). Default: tsne}

\item{...}{Extra parameters to pass to the multicoreTSNE function.}
}
\description{
Perform tSNE projection on a Seurat object using the
Multicore-opt-SNE function
}
