% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceFunction.R
\name{DooptSNE}
\alias{DooptSNE}
\title{DooptSNE}
\usage{
DooptSNE(object, reduction_use = "pca", reduction_save = "optsne",
  dims_use, ...)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object to be transformed.}

\item{reduction_use}{Prior dimensional reduction to use for calculations
(i.e. pca, ica, cca, etc...). Default: pca}

\item{reduction_save}{Name to use for the reduction (i. e. tsne, umap,
etc...). Default: tsne}

\item{dims_use}{Dimensions from `reduction_use` to pass to PHATE}

\item{...}{Extra parameters to pass to the multicoreTSNE function.}
}
\description{
Perform tSNE projection on a Seurat object using the
Multicore-opt-SNE function
}
