% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{DoUMAP}
\alias{DoUMAP}
\title{DoUMAP}
\usage{
DoUMAP(object, reduction.use = "pca", reduction.save = "umap", ...)
}
\arguments{
\item{reduction.use}{Prior dimensional reduction to use for calculations
(i.e. pca, ica, cca, etc...). Default: pca}

\item{reduction.save}{Name to use for the reduction (i. e. tsne, umap,
etc...). Default: umap}

\item{...}{Extra parameters to pass to the umap function.}
}
\description{
Perform UMAP dimentional reduction
}
