% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceFunction.R
\name{PushData}
\alias{PushData}
\alias{PushData.Seurat}
\alias{PushData.seurat}
\alias{PushData.SingleCellExperiment}
\title{PushData}
\usage{
PushData(object, ...)

\method{PushData}{Seurat}(object, python_df, reduction.save, assay.used)

\method{PushData}{seurat}(object, python_df, reduction.save)

\method{PushData}{SingleCellExperiment}(object, python_df, reduction.save)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object to be transformed.}

\item{...}{Arguments passed to specific downstream methods}

\item{python_df}{Dataframe returned by a Python function}

\item{reduction.save}{Name to use for the reduction (i. e. tsne, umap,
etc...)}

\item{assay.used}{Assay from which the data that is dimensionally reduced comes}
}
\description{
Helper function to assist entering dimensional reduction data from Python
reduction methods
}
